PhotoRec: force to uint64_t some calculated_file_size

+++ src/file_abr.c
@@ -75,8 +75,8 @@ static int header_check_abr(const unsigned char *buffer, const unsigned int buff
   }
   reset_file_recovery(file_recovery_new);
   file_recovery_new->extension=file_hint_abr.extension;
-  file_recovery_new->min_filesize=4+12+be32(hdr->size);
-  file_recovery_new->calculated_file_size=4+12+be32(hdr->size);
+  file_recovery_new->min_filesize=(uint64_t)4+12+be32(hdr->size);
+  file_recovery_new->calculated_file_size=(uint64_t)4+12+be32(hdr->size);
   if(file_recovery_new->blocksize < 12)
     return 1;
   file_recovery_new->data_check=&data_check_abr;
